# docs/swagger.yaml
definitions:
  makerchecker.UpdateMakerchecker:
    properties:
      id:
        type: string
      status:
        type: string
    required:
    - id
    - status
    type: object
  makerchecker.VerifyMaker:
    properties:
      endpoint:
        type: string
    required:
    - endpoint
    type: object
  models.HttpError:
    properties:
      code:
        example: 400
        type: integer
      data:
        additionalProperties: true
        type: object
      message:
        example: status bad request
        type: string
    type: object
  models.Makerchecker:
    properties:
      _id:
        type: string
      checkerEmail:
        type: string
      checkerId:
        type: string
      data:
        additionalProperties: true
        type: object
      endpoint:
        type: string
      makerEmail:
        type: string
      makerId:
        type: string
      status:
        type: string
    required:
    - checkerId
    - data
    - endpoint
    type: object
...
info:
  contact: {}
paths:
...
  /record:
    get:
      description: Retrieves a list of Makerchecker records
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.Makerchecker'
            type: array
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.HttpError'
      summary: Get all Makerchecker Record
      tags:
      - makerchecker
    post:
      description: Create a Makerchecker
      parameters:
      - description: makerchecker
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/models.Makerchecker'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Makerchecker'
        "400":
          description: Bad request due to invalid JSON body
          schema:
            $ref: '#/definitions/models.HttpError'
        "403":
          description: Not enough permission to do makerchecker
          schema:
            $ref: '#/definitions/models.HttpError'
        "404":
          description: Unable to find resources
          schema:
            $ref: '#/definitions/models.HttpError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.HttpError'
      summary: Create a Makerchecker
      tags:
      - makerchecker
    put:
      description: Update Makerchecker by approving or rejecting request
      parameters:
      - description: Request Body
        in: body
        name: requestBody
        required: true
        schema:
          $ref: '#/definitions/makerchecker.UpdateMakerchecker'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Makerchecker'
        "400":
          description: Bad request due to invalid JSON body
          schema:
            $ref: '#/definitions/models.HttpError'
        "403":
          description: User is not authorize to approve the request
          schema:
            $ref: '#/definitions/models.HttpError'
        "404":
          description: No makerchecker record found with makercheckerId
          schema:
            $ref: '#/definitions/models.HttpError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.HttpError'
      summary: Update Makerchecker by approving or rejecting request
      tags:
      - makerchecker
  /record/{id}:
    get:
      description: Retrieve a Makerchecker By ID
      parameters:
      - description: id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Makerchecker'
        "400":
          description: Id cannot be empty
          schema:
            $ref: '#/definitions/models.HttpError'
        "404":
          description: Record not found with Id
          schema:
            $ref: '#/definitions/models.HttpError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.HttpError'
      summary: Get Makerchecker recordby Id
      tags:
      - makerchecker
  /record/pending-approve:
    get:
      description: Retrieves a list of pending approval records
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.Makerchecker'
            type: array
        "400":
          description: Maker Id cannot be found in the header provided
          schema:
            $ref: '#/definitions/models.HttpError'
        "404":
          description: No pending requests found
          schema:
            $ref: '#/definitions/models.HttpError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.HttpError'
      summary: Get all Pending Approval as a Maker using MakerID
      tags:
      - makerchecker
  /record/to-approve:
    get:
      description: Retrieves a list of pending approval records
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.Makerchecker'
            type: array
        "400":
          description: Checker Id cannot be found in the header provided
          schema:
            $ref: '#/definitions/models.HttpError'
        "404":
          description: No pending requests found
          schema:
            $ref: '#/definitions/models.HttpError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.HttpError'
      summary: Get all Pending Approval as a Checker using CheckerID
      tags:
      - makerchecker
  /verify:
    post:
      description: Verify if a User can do Makerchecker
      parameters:
      - description: Request Body
        in: body
        name: requestBody
        required: true
        schema:
          $ref: '#/definitions/makerchecker.VerifyMaker'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad request due to invalid JSON body
          schema:
            $ref: '#/definitions/models.HttpError'
        "403":
          description: Not enough permission to do makerchecker
          schema:
            $ref: '#/definitions/models.HttpError'
        "404":
          description: Unable to find makerchecker permission
          schema:
            $ref: '#/definitions/models.HttpError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.HttpError'
      summary: Verify if a User can do Makerchecker
      tags:
      - makerchecker
swagger: "2.0"